+++ smart-maintenance-saas/tests/unit/agents/core/test_anomaly_detection_agent.py
@@ -1,268 +1,193 @@
-"""Unit tests for AnomalyDetectionAgent."""
-
-import math # Added math import
-import uuid
-from datetime import datetime
-from unittest.mock import MagicMock, patch
-
-import numpy as np
 import pytest
+from unittest.mock import Mock, patch
+import uuid # For correlation_id testing
 
 from apps.agents.core.anomaly_detection_agent import AnomalyDetectionAgent
-from core.events.event_bus import EventBus
-from data.schemas import SensorReading, SensorType
+from data.schemas import AnomalyType, SensorReading
+from core.events.event_models import DataProcessedEvent
 
 
-class TestAnomalyDetectionAgent:
-    """Unit tests for the AnomalyDetectionAgent."""
+# Mock EventBus for agent initialization
+@pytest.fixture
+def mock_event_bus():
+    return Mock()
 
-    @pytest.fixture
-    def mock_event_bus(self):
-        """Create a mock EventBus for testing."""
-        return MagicMock(spec=EventBus)
+@pytest.fixture
+def anomaly_detection_agent(mock_event_bus):
+    # Provide minimal specific_settings if necessary for initialization
+    # For _calculate_ensemble_metrics, specific_settings might not be directly needed
+    # but the agent constructor requires it.
+    agent = AnomalyDetectionAgent(agent_id="test_anomaly_agent", event_bus=mock_event_bus, specific_settings={
+        'statistical_detector_config': {}, # Default empty config
+        'default_historical_std': 0.1
+    })
+    return agent
 
-    @pytest.fixture
-    def agent(self, mock_event_bus):
-        """Create an AnomalyDetectionAgent instance for testing."""
-        return AnomalyDetectionAgent(
-            agent_id="test_agent",
-            event_bus=mock_event_bus
+class TestAnomalyDetectionAgentCalculateEnsembleMetrics:
+
+    def test_both_anomaly_z_score(self, anomaly_detection_agent):
+        is_anomaly, confidence, desc = anomaly_detection_agent._calculate_ensemble_metrics(
+            if_prediction=-1, if_score=-0.2,  # Anomaly from IF
+            stat_is_anomaly=True, stat_confidence=0.8, stat_desc="z_score_violation"
         )
-
-    @pytest.fixture
-    def sample_sensor_reading(self):
-        """Create a sample sensor reading for testing."""
-        return SensorReading(
-            sensor_id="sensor_temp_001",
-            value=25.3,
-            timestamp=datetime.utcnow(),
-            sensor_type=SensorType.TEMPERATURE,
-            unit="celsius",
-            quality=1.0,
-            correlation_id=uuid.uuid4(),
-            metadata={"location": "room_A", "building": "main"}
-        )
-
-    def test_extract_features_basic(self, agent, sample_sensor_reading):
-        """Test that _extract_features returns proper format for basic sensor reading."""
-        features = agent._extract_features(sample_sensor_reading)
-        
-        # Check return type and shape
-        assert isinstance(features, np.ndarray)
-        assert features.shape == (1, 1)  # 2D array with single feature
-        assert features[0, 0] == sample_sensor_reading.value
-
-    def test_extract_features_different_values(self, agent):
-        """Test _extract_features with different sensor values."""
-        test_values = [0.0, -10.5, 100.7, 999.99]
-        
-        for value in test_values:
-            reading = SensorReading(
-                sensor_id="test_sensor",
-                value=value,
-                timestamp=datetime.utcnow(),
-                sensor_type=SensorType.TEMPERATURE,
-                unit="celsius",
-                quality=1.0,
-                correlation_id=uuid.uuid4(),
-                metadata={}
-            )
-            
-            features = agent._extract_features(reading)
-            
-            assert isinstance(features, np.ndarray)
-            assert features.shape == (1, 1)
-            assert features[0, 0] == value
-
-    def test_extract_features_logging(self, agent, sample_sensor_reading):
-        """Test that _extract_features works without logging (simple method)."""
-        features = agent._extract_features(sample_sensor_reading)
-        
-        # Verify features are extracted correctly
-        assert features.shape == (1, 1)
-        assert features[0, 0] == sample_sensor_reading.value
-
-    def test_extract_features_with_extreme_values(self, agent):
-        """Test _extract_features with extreme values."""
-        extreme_values = [float('inf'), -float('inf'), 1e-10, 1e10]
-        
-        for value in extreme_values:
-            if not np.isfinite(value):
-                continue  # Skip infinite values for this test
-                
-            reading = SensorReading(
-                sensor_id="extreme_sensor",
-                value=value,
-                timestamp=datetime.utcnow(),
-                sensor_type=SensorType.PRESSURE,
-                unit="pa",
-                quality=1.0,
-                correlation_id=uuid.uuid4(),
-                metadata={}
-            )
-            
-            features = agent._extract_features(reading)
-            
-            assert isinstance(features, np.ndarray)
-            assert features.shape == (1, 1)
-            assert features[0, 0] == value
-
-    def test_extract_features_multiple_calls_consistency(self, agent, sample_sensor_reading):
-        """Test that multiple calls to _extract_features return consistent results."""
-        features1 = agent._extract_features(sample_sensor_reading)
-        features2 = agent._extract_features(sample_sensor_reading)
-        
-        # Results should be identical
-        np.testing.assert_array_equal(features1, features2)
-
-    def test_extract_features_different_sensor_types(self, agent):
-        """Test _extract_features with different sensor types."""
-        sensor_configs = [
-            {"type": SensorType.TEMPERATURE, "value": 22.5, "unit": "celsius"},
-            {"type": SensorType.VIBRATION, "value": 0.05, "unit": "m/s2"},
-            {"type": SensorType.PRESSURE, "value": 101.3, "unit": "kpa"},
-        ]
-        
-        for config in sensor_configs:
-            reading = SensorReading(
-                sensor_id=f"sensor_{config['type']}_001",
-                value=config["value"],
-                timestamp=datetime.utcnow(),
-                sensor_type=config["type"],
-                unit=config["unit"],
-                quality=1.0,
-                correlation_id=uuid.uuid4(),
-                metadata={}
-            )
-            
-            features = agent._extract_features(reading)
-            
-            assert isinstance(features, np.ndarray)
-            assert features.shape == (1, 1)
-            assert features[0, 0] == config["value"]
-
-    def test_extract_features_returns_sklearn_compatible_format(self, agent, sample_sensor_reading):
-        """Test that _extract_features returns format compatible with sklearn."""
-        features = agent._extract_features(sample_sensor_reading)
-        
-        # Should be 2D array (required by sklearn)
-        assert len(features.shape) == 2
-        assert features.shape[0] >= 1  # At least one sample
-        assert features.shape[1] >= 1  # At least one feature
-        
-        # Should be numeric type compatible with sklearn
-        assert np.issubdtype(features.dtype, np.number)
-
-    def test_ensemble_decision_both_anomaly(self, agent):
-        """Test ensemble decision when both methods detect anomaly."""
-        is_anomaly, confidence, description = agent._ensemble_decision(
-            if_prediction=-1,      # IF detects anomaly
-            if_score=-0.2,         # Negative score indicates anomaly
-            stat_is_anomaly=True,  # Statistical method detects anomaly
-            stat_confidence=0.8,   # High statistical confidence
-            stat_desc="threshold_breach"
-        )
-        
         assert is_anomaly is True
-        assert confidence > 0.5  # Should have reasonable confidence
-        assert "ensemble_anomaly_if_and_threshold_breach" in description
+        assert 0.5 < confidence <= 1.0 # Combined confidence
+        # Expected: (0.6 * (0.5 + abs(-0.2)*0.5) + 0.4 * 0.8) = (0.6 * 0.6 + 0.32) = 0.36 + 0.32 = 0.68
+        assert abs(confidence - 0.68) < 0.001
+        assert desc == AnomalyType.ENSEMBLE_IF_STATISTICAL
 
-    def test_ensemble_decision_only_isolation_forest(self, agent):
-        """Test ensemble decision when only Isolation Forest detects anomaly."""
-        is_anomaly, confidence, description = agent._ensemble_decision(
-            if_prediction=-1,      # IF detects anomaly
-            if_score=-0.3,         # Strong anomaly signal
-            stat_is_anomaly=False, # Statistical method says normal
-            stat_confidence=0.1,   # Low statistical confidence
-            stat_desc="normal"
+    def test_both_anomaly_threshold(self, anomaly_detection_agent):
+        is_anomaly, confidence, desc = anomaly_detection_agent._calculate_ensemble_metrics(
+            if_prediction=-1, if_score=-0.1,
+            stat_is_anomaly=True, stat_confidence=0.7, stat_desc="threshold_violation"
         )
-        
         assert is_anomaly is True
-        assert confidence > 0.0
-        assert description == "isolation_forest_anomaly"
+        # Expected: (0.6 * (0.5 + abs(-0.1)*0.5) + 0.4 * 0.7) = (0.6 * 0.55 + 0.28) = 0.33 + 0.28 = 0.61
+        assert abs(confidence - 0.61) < 0.001
+        assert desc == AnomalyType.ENSEMBLE_IF_STATISTICAL
 
-    def test_ensemble_decision_only_statistical(self, agent):
-        """Test ensemble decision when only statistical method detects anomaly."""
-        is_anomaly, confidence, description = agent._ensemble_decision(
-            if_prediction=1,       # IF says normal
-            if_score=0.1,          # Positive score indicates normal
-            stat_is_anomaly=True,  # Statistical method detects anomaly
-            stat_confidence=0.9,   # High statistical confidence
-            stat_desc="spike_detected"
+    def test_both_anomaly_other_stat(self, anomaly_detection_agent):
+        is_anomaly, confidence, desc = anomaly_detection_agent._calculate_ensemble_metrics(
+            if_prediction=-1, if_score=-0.1,
+            stat_is_anomaly=True, stat_confidence=0.7, stat_desc="other_violation"
         )
-        
         assert is_anomaly is True
-        assert math.isclose(confidence, 0.72)  # Should use statistical confidence * 0.8
-        assert description == "statistical_spike_detected" # Code generates lowercase 'statistical_'
+        assert abs(confidence - 0.61) < 0.001 # Same confidence as above
+        assert desc == AnomalyType.ENSEMBLE_IF_STATISTICAL # Falls back to general ensemble
 
-    def test_ensemble_decision_no_anomaly(self, agent):
-        """Test ensemble decision when neither method detects anomaly."""
-        is_anomaly, confidence, description = agent._ensemble_decision(
-            if_prediction=1,       # IF says normal
-            if_score=0.2,          # Positive score indicates normal
-            stat_is_anomaly=False, # Statistical method says normal
-            stat_confidence=0.0,   # No statistical confidence
-            stat_desc="normal"
+    def test_if_only_anomaly(self, anomaly_detection_agent):
+        is_anomaly, confidence, desc = anomaly_detection_agent._calculate_ensemble_metrics(
+            if_prediction=-1, if_score=-0.3,
+            stat_is_anomaly=False, stat_confidence=0.2, stat_desc="normal"
         )
-        
+        assert is_anomaly is True
+        # Expected: (0.5 + abs(-0.3)*0.5) * 0.8 = (0.5 + 0.15) * 0.8 = 0.65 * 0.8 = 0.52
+        assert abs(confidence - 0.52) < 0.001
+        assert desc == AnomalyType.ISOLATION_FOREST
+
+    def test_stat_only_anomaly_z_score(self, anomaly_detection_agent):
+        is_anomaly, confidence, desc = anomaly_detection_agent._calculate_ensemble_metrics(
+            if_prediction=1, if_score=0.1, # Normal from IF
+            stat_is_anomaly=True, stat_confidence=0.9, stat_desc="z_score_violation"
+        )
+        assert is_anomaly is True
+        # Expected: 0.9 * 0.8 = 0.72
+        assert abs(confidence - 0.72) < 0.001
+        assert desc == AnomalyType.STATISTICAL_Z_SCORE
+
+    def test_stat_only_anomaly_threshold(self, anomaly_detection_agent):
+        is_anomaly, confidence, desc = anomaly_detection_agent._calculate_ensemble_metrics(
+            if_prediction=1, if_score=0.2,
+            stat_is_anomaly=True, stat_confidence=0.75, stat_desc="threshold_violation"
+        )
+        assert is_anomaly is True
+        # Expected: 0.75 * 0.8 = 0.6
+        assert abs(confidence - 0.6) < 0.001
+        assert desc == AnomalyType.STATISTICAL_THRESHOLD
+
+    def test_stat_only_anomaly_unknown_desc(self, anomaly_detection_agent):
+        is_anomaly, confidence, desc = anomaly_detection_agent._calculate_ensemble_metrics(
+            if_prediction=1, if_score=0.2,
+            stat_is_anomaly=True, stat_confidence=0.75, stat_desc="some_other_stat_finding"
+        )
+        assert is_anomaly is True
+        assert abs(confidence - 0.6) < 0.001
+        assert desc == AnomalyType.UNKNOWN # As per current logic
+
+    def test_no_anomaly(self, anomaly_detection_agent):
+        is_anomaly, confidence, desc = anomaly_detection_agent._calculate_ensemble_metrics(
+            if_prediction=1, if_score=0.5, # Normal from IF
+            stat_is_anomaly=False, stat_confidence=0.1, stat_desc="normal"
+        )
         assert is_anomaly is False
-        assert confidence == 0.0
-        assert description == "normal"
+        assert confidence == 0.0 # Based on current logic where final_confidence_score is initialized to 0.0
+        assert desc == "normal"
 
-    def test_ensemble_decision_confidence_mapping(self, agent):
-        """Test that Isolation Forest scores are properly mapped to confidence."""
-        # Test high anomaly score
-        is_anomaly, confidence, description = agent._ensemble_decision(
-            if_prediction=-1,
-            if_score=-0.5,         # Very negative score
-            stat_is_anomaly=False,
-            stat_confidence=0.0,
-            stat_desc="normal"
+    def test_if_score_normalization_edge_cases(self, anomaly_detection_agent):
+        # if_score at max anomaly (-1.0 for normalization cap)
+        _, confidence_max_neg_if, _ = anomaly_detection_agent._calculate_ensemble_metrics(
+            if_prediction=-1, if_score=-1.0, stat_is_anomaly=False, stat_confidence=0.1, stat_desc="normal"
         )
-        
-        assert is_anomaly is True
-        assert confidence >= 0.5  # Should have at least minimum confidence
-        
-        # Test moderate anomaly score
-        is_anomaly2, confidence2, description2 = agent._ensemble_decision(
-            if_prediction=-1,
-            if_score=-0.1,         # Less negative score
-            stat_is_anomaly=False,
-            stat_confidence=0.0,
-            stat_desc="normal"
-        )
-        
-        assert is_anomaly2 is True
-        assert confidence2 >= 0.5  # Should still have minimum confidence
+        # Expected if_confidence: (0.5 + abs(-1.0)*0.5) = 1.0. Then 1.0 * 0.8 = 0.8
+        assert abs(confidence_max_neg_if - 0.8) < 0.001
 
-    def test_ensemble_decision_max_confidence(self, agent):
-        """Test that ensemble takes maximum of individual confidences."""
-        # Statistical confidence higher
-        is_anomaly, confidence, description = agent._ensemble_decision(
-            if_prediction=-1,
-            if_score=-0.1,         # Moderate IF confidence
-            stat_is_anomaly=True,
-            stat_confidence=0.95,  # Very high statistical confidence
-            stat_desc="outlier"
+        # if_score close to 0 but still anomaly
+        _, confidence_min_neg_if, _ = anomaly_detection_agent._calculate_ensemble_metrics(
+            if_prediction=-1, if_score=-0.01, stat_is_anomaly=False, stat_confidence=0.1, stat_desc="normal"
         )
-        
-        assert is_anomaly is True
-        assert math.isclose(confidence, 0.845)  # Corrected expected confidence
+        # Expected if_confidence: (0.5 + abs(-0.01)*0.5) = 0.505. Then 0.505 * 0.8 = 0.404
+        assert abs(confidence_min_neg_if - 0.404) < 0.001
 
-    def test_ensemble_decision_logging(self, agent):
-        """Test that ensemble decision logs debug information."""
-        with patch.object(agent.logger, 'debug') as mock_debug:
-            agent._ensemble_decision(
-                if_prediction=-1,
-                if_score=-0.2,
-                stat_is_anomaly=True,
-                stat_confidence=0.7,
-                stat_desc="test_anomaly"
-            )
-            
-            # Verify debug logging was called
-            mock_debug.assert_called_once()
-            log_message = mock_debug.call_args[0][0]
-            assert "Ensemble:" in log_message  # Match actual log format
-            assert "final_anom=True" in log_message # Corrected key for anomaly status
-            assert "final_conf=" in log_message # Corrected key for confidence
+        # if_score at 0 (should be treated as anomaly by prediction, but score shows no confidence)
+        _, confidence_zero_if, _ = anomaly_detection_agent._calculate_ensemble_metrics(
+            if_prediction=-1, if_score=0.0, stat_is_anomaly=False, stat_confidence=0.1, stat_desc="normal"
+        )
+        # Expected if_confidence: (0.5 + abs(0.0)*0.5) = 0.5. Then 0.5 * 0.8 = 0.4
+        assert abs(confidence_zero_if - 0.4) < 0.001
+
+# More tests will be added for logging correlation_id
+# and potentially for the process method if needed.
+# For now, focusing on the refactored method.
+
+# Test for correlation_id in logging
+@patch.object(AnomalyDetectionAgent, '_process_ml_models', return_value=(1, 0.1)) # Mock to prevent actual ML processing
+@patch.object(AnomalyDetectionAgent, '_process_statistical_method', return_value=(False, 0.1, "normal"))
+@patch.object(AnomalyDetectionAgent, '_handle_anomaly_detected') # We are not testing this part here
+async def test_process_logging_with_correlation_id(
+    mock_handle_anomaly,
+    mock_process_stat,
+    mock_process_ml,
+    anomaly_detection_agent, # use the fixture
+    mocker
+):
+    # Patch the logger directly on the agent instance
+    mock_logger = mocker.patch.object(anomaly_detection_agent, 'logger')
+
+    correlation_id = str(uuid.uuid4())
+    event_data = {
+        "event_id": str(uuid.uuid4()),
+        "timestamp": "2023-10-26T10:00:00Z",
+        "correlation_id": correlation_id,
+        "processed_data": {
+            "sensor_id": "sensor123",
+            "value": 100.0,
+            "timestamp": "2023-10-26T10:00:00Z", # SensorReading needs this
+            "sensor_type": "temperature", # SensorReading needs this
+            "unit": "C" # SensorReading needs this
+        }
+    }
+    event = DataProcessedEvent(**event_data)
+
+    # Setup scaler and isolation_forest_fitted attributes that are accessed in process
+    anomaly_detection_agent.scaler = Mock()
+    anomaly_detection_agent.scaler.fit_transform = Mock(return_value=([[100.0]]))
+    anomaly_detection_agent.isolation_forest_fitted = True # Assume fitted for this test path
+
+
+    await anomaly_detection_agent.process(event)
+
+    # Check if logger.debug was called with correlation_id in extra
+    found_log_with_correlation_id = False
+    for call in mock_logger.debug.call_args_list:
+        args, kwargs = call
+        if kwargs.get('extra') == {"correlation_id": correlation_id}:
+            found_log_with_correlation_id = True
+            break
+    assert found_log_with_correlation_id, "logger.debug was not called with correct correlation_id in extra"
+
+    # Also check logger.info for the "Final decision" log
+    found_final_decision_log = False
+    for call_args in mock_logger.info.call_args_list:
+         if "Final decision for sensor123" in call_args[0][0] and \
+            call_args[1].get('extra') == {"correlation_id": correlation_id}:
+            found_final_decision_log = True
+            break
+    assert found_final_decision_log, "Final decision log was not called with correct correlation_id"
+
+    # If an anomaly was detected and _handle_anomaly_detected was called,
+    # we'd also check logs from there if they are direct calls from AnomalyDetectionAgent.
+    # In this case, mock_handle_anomaly is used, so we don't check its internal logs here.
+
+    # The process method also logs errors. A separate test could be made for error paths.
+    # For now, this covers a non-error path.
+pytest
