--- tests/unit/agents/decision/test_scheduling_agent.py
@@ -1,308 +1,0 @@
-import pytest
-from unittest.mock import Mock, patch
-from datetime import datetime, timedelta
-import uuid
-
-from smart_maintenance_saas.apps.agents.decision.scheduling_agent import SchedulingAgent, CalendarService
-from smart_maintenance_saas.data.schemas import MaintenanceRequest, ScheduleStatus, OptimizedSchedule
-from smart_maintenance_saas.core.events.event_models import MaintenancePredictedEvent
-
-
-# Mock EventBus for agent initialization
-@pytest.fixture
-def mock_event_bus():
-    return Mock()
-
-@pytest.fixture
-def scheduling_agent(mock_event_bus):
-    agent = SchedulingAgent(agent_id="test_scheduling_agent", event_bus=mock_event_bus)
-    return agent
-
-# Mock technician data for tests
-MOCK_TECHNICIANS = [
-    {"id": "tech_001", "name": "John Smith", "skills": ["electrical", "mechanical"], "experience_years": 5, "availability_score": 0.8},
-    {"id": "tech_002", "name": "Sarah Johnson", "skills": ["mechanical", "plumbing"], "experience_years": 8, "availability_score": 0.9},
-    {"id": "tech_003", "name": "Mike Brown", "skills": ["electrical"], "experience_years": 3, "availability_score": 0.95}, # Higher availability, less experience
-    {"id": "tech_004", "name": "Emily White", "skills": ["plumbing", "hvac"], "experience_years": 10, "availability_score": 0.7},
-]
-
-class TestSchedulingAgentFindBestQualifiedTechnician:
-
-    def test_find_best_tech_all_skills_match(self, scheduling_agent):
-        request = MaintenanceRequest(
-            id="req1", equipment_id="eq1", maintenance_type="repair", priority=1,
-            estimated_duration_hours=2, required_skills=["electrical", "mechanical"]
-        )
-        # Expected: tech_001 (John Smith) due to skill match and sorting (0.8 avail, 5 exp)
-        # tech_002 has mechanical but not electrical.
-        best_tech = scheduling_agent._find_best_qualified_technician(request, MOCK_TECHNICIANS)
-        assert best_tech is not None
-        assert best_tech["id"] == "tech_001"
-
-    def test_find_best_tech_partial_skill_match_highest_score(self, scheduling_agent):
-        request = MaintenanceRequest(
-            id="req2", equipment_id="eq2", maintenance_type="inspection", priority=2,
-            estimated_duration_hours=1, required_skills=["mechanical"]
-        )
-        # Expected: tech_002 (Sarah Johnson) - mechanical, 0.9 avail, 8 exp
-        # tech_001 also has mechanical, but 0.8 avail, 5 exp
-        best_tech = scheduling_agent._find_best_qualified_technician(request, MOCK_TECHNICIANS)
-        assert best_tech is not None
-        assert best_tech["id"] == "tech_002"
-
-    def test_find_best_tech_sorting_logic(self, scheduling_agent):
-        # Techs with 'electrical' skill:
-        # tech_001: avail=0.8, exp=5
-        # tech_003: avail=0.95, exp=3
-        # Expected: tech_003 due to higher availability_score, despite less experience.
-        request = MaintenanceRequest(
-            id="req3", equipment_id="eq3", maintenance_type="repair", priority=1,
-            estimated_duration_hours=3, required_skills=["electrical"]
-        )
-        best_tech = scheduling_agent._find_best_qualified_technician(request, MOCK_TECHNICIANS)
-        assert best_tech is not None
-        assert best_tech["id"] == "tech_003"
-
-    def test_no_technician_with_required_skills(self, scheduling_agent):
-        request = MaintenanceRequest(
-            id="req4", equipment_id="eq4", maintenance_type="repair", priority=1,
-            estimated_duration_hours=2, required_skills=["robotics"] # No tech has robotics
-        )
-        best_tech = scheduling_agent._find_best_qualified_technician(request, MOCK_TECHNICIANS)
-        assert best_tech is None
-
-    def test_empty_technician_list(self, scheduling_agent):
-        request = MaintenanceRequest(
-            id="req5", equipment_id="eq5", maintenance_type="repair", priority=1,
-            estimated_duration_hours=2, required_skills=["mechanical"]
-        )
-        best_tech = scheduling_agent._find_best_qualified_technician(request, [])
-        assert best_tech is None
-
-    def test_request_with_no_specific_skills(self, scheduling_agent):
-        # If no skills are required, the current logic `any(skill in tech["skills"] for skill in maintenance_request.required_skills)`
-        # will result in no technicians if required_skills is empty, because `any` on an empty list is False.
-        # This might be desired (treat as "no one is qualified if no skills listed") or might need adjustment
-        # in the main method to assign a default skill if empty.
-        # For this test, we assume current logic: empty required_skills means no one matches.
-        request = MaintenanceRequest(
-            id="req6", equipment_id="eq6", maintenance_type="checkup", priority=3,
-            estimated_duration_hours=1, required_skills=[]
-        )
-        # Based on current implementation of _find_best_qualified_technician,
-        # if maintenance_request.required_skills is empty, `any(...)` will be False.
-        best_tech = scheduling_agent._find_best_qualified_technician(request, MOCK_TECHNICIANS)
-        assert best_tech is None # Or assert it returns the top sorted tech if logic changes
-                                 # Current logic: any() on empty list is False, so filters out everyone.
-                                 # If the intent is "any tech is fine if no skills specified", the filter needs adjustment.
-                                 # For testing, we test the current implemented behavior.
-
-
-@pytest.fixture
-def mock_calendar_service():
-    service = Mock(spec=CalendarService)
-    # Configure default return values for mocked methods
-    service.check_availability.return_value = True
-    service.book_slot.return_value = True
-    return service
-
-class TestSchedulingAgentFindAvailableSlot:
-
-    @pytest.mark.freeze_time('2023-10-26 12:00:00') # Freeze time for consistent `datetime.utcnow()`
-    def test_slot_found_and_booked(self, scheduling_agent, mock_calendar_service):
-        scheduling_agent.calendar_service = mock_calendar_service
-        technician = MOCK_TECHNICIANS[0] # John Smith
-        request = MaintenanceRequest(
-            id="req_slot1", equipment_id="eq_slot1", maintenance_type="repair", priority=1,
-            estimated_duration_hours=2, required_skills=["electrical"],
-            # Preferred window starts in the future relative to frozen time
-            preferred_time_window_start=datetime(2023, 10, 27, 9, 0, 0),
-            preferred_time_window_end=datetime(2023, 10, 27, 17, 0, 0)
-        )
-
-        expected_start = datetime(2023, 10, 27, 9, 0, 0)
-        expected_end = expected_start + timedelta(hours=request.estimated_duration_hours)
-
-        slot = scheduling_agent._find_available_slot_for_technician(technician, request)
-
-        assert slot is not None
-        assert slot[0] == expected_start
-        assert slot[1] == expected_end
-        mock_calendar_service.check_availability.assert_called_once_with(technician["id"], expected_start, expected_end)
-        mock_calendar_service.book_slot.assert_called_once_with(technician["id"], expected_start, expected_end, request.id)
-
-    @pytest.mark.freeze_time('2023-10-26 12:00:00')
-    def test_no_slot_due_to_unavailability(self, scheduling_agent, mock_calendar_service):
-        scheduling_agent.calendar_service = mock_calendar_service
-        mock_calendar_service.check_availability.return_value = False # Always unavailable
-
-        technician = MOCK_TECHNICIANS[1] # Sarah Johnson
-        request = MaintenanceRequest(
-            id="req_slot2", equipment_id="eq_slot2", maintenance_type="inspection", priority=2,
-            estimated_duration_hours=1, required_skills=["mechanical"],
-            preferred_time_window_start=datetime(2023, 10, 27, 9, 0, 0),
-            preferred_time_window_end=datetime(2023, 10, 27, 10, 0, 0) # Short window
-        )
-
-        slot = scheduling_agent._find_available_slot_for_technician(technician, request)
-
-        assert slot is None
-        mock_calendar_service.book_slot.assert_not_called()
-        # Check that check_availability was called at least once for the start of the window
-        # The exact number of calls depends on loop increment and window size.
-        # For a 1-hour window and 1-hour increment, it should be called for 9:00.
-        expected_check_start = datetime(2023, 10, 27, 9, 0, 0)
-        expected_check_end = expected_check_start + timedelta(hours=request.estimated_duration_hours)
-        mock_calendar_service.check_availability.assert_any_call(technician["id"], expected_check_start, expected_check_end)
-
-
-    @pytest.mark.freeze_time('2023-10-26 12:00:00')
-    def test_slot_found_but_booking_fails(self, scheduling_agent, mock_calendar_service):
-        scheduling_agent.calendar_service = mock_calendar_service
-        mock_calendar_service.check_availability.return_value = True
-        mock_calendar_service.book_slot.return_value = False # Booking fails
-
-        technician = MOCK_TECHNICIANS[0]
-        request = MaintenanceRequest(
-            id="req_slot3", equipment_id="eq_slot3", maintenance_type="repair", priority=1,
-            estimated_duration_hours=3, required_skills=["electrical"],
-            preferred_time_window_start=datetime(2023, 10, 27, 10, 0, 0),
-            preferred_time_window_end=datetime(2023, 10, 27, 14, 0, 0)
-        )
-
-        slot = scheduling_agent._find_available_slot_for_technician(technician, request)
-
-        assert slot is None
-        expected_check_start = datetime(2023, 10, 27, 10, 0, 0)
-        expected_check_end = expected_check_start + timedelta(hours=request.estimated_duration_hours)
-        mock_calendar_service.check_availability.assert_called_with(technician["id"], expected_check_start, expected_check_end)
-        mock_calendar_service.book_slot.assert_called_with(technician["id"], expected_check_start, expected_check_end, request.id)
-
-    @pytest.mark.freeze_time('2023-10-28 10:00:00') # Saturday
-    def test_slot_search_skips_weekend_and_finds_monday(self, scheduling_agent, mock_calendar_service):
-        scheduling_agent.calendar_service = mock_calendar_service
-        technician = MOCK_TECHNICIANS[0]
-        request = MaintenanceRequest(
-            id="req_slot_weekend", equipment_id="eq_slot_weekend", maintenance_type="repair", priority=1,
-            estimated_duration_hours=2, required_skills=["electrical"],
-            # Preferred window starts on Saturday, should find slot on Monday
-            preferred_time_window_start=datetime(2023, 10, 28, 9, 0, 0),
-            preferred_time_window_end=datetime(2023, 10, 30, 17, 0, 0)
-        )
-
-        # Calendar service should be set to True for the expected Monday slot
-        # It will be called for Saturday/Sunday and then Monday.
-        # We want to test that the logic correctly advances to Monday.
-
-        expected_monday_start = datetime(2023, 10, 30, 8, 0, 0) # Adjusted to 8 AM by the method
-        expected_monday_end = expected_monday_start + timedelta(hours=request.estimated_duration_hours)
-
-        def side_effect_check_availability(tech_id, start_time, end_time):
-            if start_time.weekday() < 5 and start_time.hour >= 8 and end_time.hour <=18: # Monday-Friday, business hours
-                 if start_time == expected_monday_start : return True
-            return False
-
-        mock_calendar_service.check_availability.side_effect = side_effect_check_availability
-
-        slot = scheduling_agent._find_available_slot_for_technician(technician, request)
-
-        assert slot is not None
-        assert slot[0] == expected_monday_start
-        assert slot[1] == expected_monday_end
-        mock_calendar_service.book_slot.assert_called_once_with(technician["id"], expected_monday_start, expected_monday_end, request.id)
-
-    @pytest.mark.freeze_time('2023-10-26 17:00:00') # End of day
-    def test_slot_search_starts_next_day_if_too_late(self, scheduling_agent, mock_calendar_service):
-        scheduling_agent.calendar_service = mock_calendar_service
-        technician = MOCK_TECHNICIANS[0]
-        request = MaintenanceRequest(
-            id="req_slot_late", equipment_id="eq_slot_late", maintenance_type="repair", priority=1,
-            estimated_duration_hours=4, required_skills=["electrical"], # 4-hour task
-            # Preferred window starts now (17:00), too late for a 4-hour task.
-            # Max start hour would be 18 - 4 = 14:00.
-            preferred_time_window_start=datetime(2023, 10, 26, 17, 0, 0),
-            preferred_time_window_end=datetime(2023, 10, 27, 17, 0, 0)
-        )
-
-        expected_next_day_start = datetime(2023, 10, 27, 8, 0, 0)
-        expected_next_day_end = expected_next_day_start + timedelta(hours=request.estimated_duration_hours)
-
-        slot = scheduling_agent._find_available_slot_for_technician(technician, request)
-
-        assert slot is not None
-        assert slot[0] == expected_next_day_start
-        assert slot[1] == expected_next_day_end
-        mock_calendar_service.check_availability.assert_called_with(technician["id"], expected_next_day_start, expected_next_day_end)
-        mock_calendar_service.book_slot.assert_called_with(technician["id"], expected_next_day_start, expected_next_day_end, request.id)
-
-# Tests for correlation_id logging in SchedulingAgent
-@patch('smart_maintenance_saas.apps.agents.decision.scheduling_agent.logging')
-async def test_handle_event_logging_with_correlation_id(
-    mock_logging,
-    scheduling_agent # Use the fixture
-):
-    mock_logger = Mock()
-    mock_logging.getLogger.return_value = mock_logger # Agent's self.logger will be this mock
-
-    correlation_id = str(uuid.uuid4())
-    event_data = {
-        "event_id": str(uuid.uuid4()),
-        "timestamp": "2023-10-26T10:00:00Z",
-        "equipment_id": "eq_corr_test",
-        "predicted_failure_date": "2023-11-25T10:00:00Z",
-        "time_to_failure_days": 30,
-        "maintenance_type": "preventive",
-        "confidence": 0.85,
-        "reason": "Pattern detected",
-        "model_id": "model_v2",
-        "version": "1.0",
-        "correlation_id": correlation_id
-    }
-    event = MaintenancePredictedEvent(**event_data)
-
-    # Mock dependent methods to isolate handle_maintenance_predicted_event's logging
-    scheduling_agent._create_maintenance_request = Mock(return_value=MaintenanceRequest(
-        id="mock_req", equipment_id="eq_corr_test", maintenance_type="preventive", priority=2, estimated_duration_hours=2
-    ))
-
-    # Mock schedule_maintenance_task to return a "failed to schedule" to simplify
-    scheduling_agent.schedule_maintenance_task = Mock(
-        return_value=OptimizedSchedule(
-            request_id="mock_req", status=ScheduleStatus.FAILED_TO_SCHEDULE
-        )
-    )
-    # If it were SCHEDULED, it would try to publish, which might need more mocks.
-
-    await scheduling_agent.handle_maintenance_predicted_event("MaintenancePredictedEvent", event)
-
-    # Check the first log call in handle_maintenance_predicted_event
-    # Example: logger.info(f"Received {event_type} for equipment {equipment_id}", extra={"correlation_id": correlation_id})
-    found_received_log = False
-    for call_args in mock_logger.info.call_args_list:
-        if f"Received MaintenancePredictedEvent for equipment {event.equipment_id}" in call_args[0][0] and \
-           call_args[1].get('extra') == {"correlation_id": correlation_id}:
-            found_received_log = True
-            break
-    assert found_received_log, "Initial log in handle_maintenance_predicted_event did not include correct correlation_id"
-
-    # Example: logger.info(f"Created maintenance request ...", extra={"correlation_id": correlation_id})
-    found_created_req_log = False
-    for call_args in mock_logger.info.call_args_list:
-        if "Created maintenance request" in call_args[0][0] and \
-           call_args[1].get('extra') == {"correlation_id": correlation_id}:
-            found_created_req_log = True
-            break
-    assert found_created_req_log, "Log for created maintenance request did not include correct correlation_id"
-
-    # Example: logger.warning(f"Failed to schedule ...", extra={"correlation_id": correlation_id})
-    found_failed_schedule_log = False
-    for call_args in mock_logger.warning.call_args_list:
-        if "Failed to schedule maintenance for request" in call_args[0][0] and \
-           call_args[1].get('extra') == {"correlation_id": correlation_id}:
-            found_failed_schedule_log = True
-            break
-    assert found_failed_schedule_log, "Log for failed schedule did not include correct correlation_id"
-
-# More tests can be added for schedule_maintenance_task itself to ensure it calls helpers correctly
-# and for _publish_maintenance_scheduled_event logging.
-pytest
